scripts/gcloud/requirements.txt
google-api-core==2.1.1
google-auth==2.3.0
google-cloud-compute==0.5.0
googleapis-common-protos==1.53.0